refactor(mesh): extract MeshTransport (finding 1 job 1; finding 19)
ESP-NOW setup, RX ring buffer + trampoline, and send primitives move
into a new MeshTransport class. registerPeerWithEspNow moves here from
MeshCrypto.h (peering, not crypto). Mesh holds one as a member and
delegates all radio I/O to it; message dispatch (proto/replay checks +
message-type switch) stays on Mesh as handleReceivedMessage, reached
via a MessageHandler passed to MeshTransport::drain().

Also fixes a real bug this move exposed: the e2e SimNode harness swaps
Mesh::instance per simulated node but didn't know about the new
MeshTransport::instance singleton, so ESP-NOW RX callbacks delivered
into whichever node's transport was constructed last instead of the
currently-swapped-in node (32/41 e2e failures until fixed). NodeContext
now swaps MeshTransport::instance alongside Mesh::instance.

269/269 unit + 41/41 e2e pass.
