+++ tests/unit/test_route_report.cpp
@@ -497,13 +497,13 @@ TEST_F(RouteReportTest, DrainRecvQueue_DispatchesRouteReport) {
 
   // Directly push to recv queue (UNIT_TEST exposes all members)
   uint8_t srcMac[6] = {0x11, 0x22, 0x33, 0x44, 0x55, 0x66};
-  Mesh::RecvQueueEntry entry;
+  MeshTransport::RecvQueueEntry entry;
   memcpy(entry.srcMac, srcMac, 6);
   entry.msg = msg;
-  xRingbufferSend(mesh.recvQueue, &entry, sizeof(entry), 0);
+  xRingbufferSend(mesh.transport.recvQueue, &entry, sizeof(entry), 0);
 
   size_t before = espNowSentPackets.size();
-  mesh.drainRecvQueue();
+  mesh.drain();
 
   EXPECT_GT(espNowSentPackets.size(), before); // relayed the message
 }
